Refactor accordions into its own component.
